refactor: improve module naming and example structure
